Fix Delete_student.java's error & Edit Add_student.java & Make Network util file and folder
