--- StudentInformationSystem/src/data/Student.java
@@ -1,8 +1,0 @@
-package data;
-
-public class Student {
-	int id;
-	String name;
-	String department;
-	String phone_number;
-}
