+++ StudentInformationSystem/src/util/NetworkUtils.java
@@ -1,0 +1,5 @@
+package util;
+
+public class NetworkUtils {
+
+}
